firefox may not support chrome.offscreen, but we can avoid funtion duplication, browserpass/browserpass-extension#320
