+++ src/offscreen/offscreen.html
@@ -1,3 +1,5 @@
 <!DOCTYPE html>
-<textarea id="text"></textarea>
-<script src="../js/offscreen.dist.js"></script>
+<html>
+    <body></body>
+    <script src="../js/offscreen.dist.js"></script>
+</html>
